requirements.txt
alembic==1.0.10
Click==7.0
elasticsearch==7.0.1
Flask==1.0.2
Flask-Elasticsearch==0.2.5
Flask-JWT==0.3.2
Flask-Login==0.4.1
Flask-Migrate==2.4.0
Flask-Moment==0.7.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.10
MarkupSafe==1.1.1
psycopg2==2.8.2
PyJWT==1.4.2
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.3
urllib3==1.25.2
Werkzeug==0.15.2
WTForms==2.2.1
